refactor: replace config-based ImageFormat with static type and move models to src/models
- Define all supported image formats as a static union type (ImageFormat)
- Move all type/model files from src/shared/models to src/models
- Update all imports and remove config-based format typing

Closes #37
