+++ src/backend/index.ts
@@ -1,4 +1,4 @@
-import { PixstoreBackendConfig } from '../shared/models/pixstore-config'
+import { PixstoreBackendConfig } from '../models/pixstore-config'
 import { initPixstore, pixstoreConfig } from '../shared/pixstore-config'
 import { startDefaultEndpoint } from './default-endpoint'
 
@@ -22,6 +22,6 @@ export {
   getImage,
 } from './image-service'
 
-export type { ImageRecord } from '../shared/models/image-record'
+export type { ImageRecord } from '../models/image-record'
 
 export { customEndpointHelper } from './custom-endpoint'
