+++ src/frontend/format-image.ts
@@ -1,6 +1,6 @@
 import { decodeImagePayload } from '../shared/image-encoder'
-import { FrontendImageRecord } from '../shared/models/frontend-image-record'
-import { ImageFormat } from '../shared/models/image-format'
+import { FrontendImageRecord } from '../models/frontend-image-record'
+import { ImageFormat } from '../models/image-format'
 import { pixstoreConfig } from '../shared/pixstore-config'
 
 const IMAGE_FORMATS = pixstoreConfig.imageFormats
