+++ src/models/frontend-image-record.ts
