+++ src/models/image-format.ts
@@ -1,0 +1,29 @@
+export type ImageFormat =
+  | 'jpeg'
+  | 'jpg'
+  | 'png'
+  | 'webp'
+  | 'gif'
+  | 'bmp'
+  | 'tiff'
+  | 'tif'
+  | 'ico'
+  | 'svg'
+  | 'avif'
+  | 'heif'
+  | 'heic'
+  | 'jfif'
+  | 'pjpeg'
+  | 'pjp'
+  | 'apng'
+  | 'raw'
+  | 'emf'
+  | 'wmf'
+  | 'cur'
+  | 'dds'
+  | 'icns'
+  | 'ras'
+  | 'sgi'
+  | 'tga'
+  | 'xbm'
+  | 'xpm'
