+++ src/models/image-payload.ts
