+++ src/models/image-record.ts
