+++ src/models/pixstore-config.ts
