--- src/shared/models/image-format.ts
@@ -1,3 +1,0 @@
-import { IMAGE_FORMATS } from '../pixstore-config'
-
-export type ImageFormat = (typeof IMAGE_FORMATS)[number]
